requirements.txt
Flask==3.0.3
singlestoredb==1.7.2
pillow==10.4.0
sentence-transformers==3.2.0
groq